+++ package.json
@@ -5,7 +5,7 @@
   "homepage": "http://github.com/wearefractal/gulp-jshint",
   "repository": "git://github.com/wearefractal/gulp-jshint.git",
   "author": "Fractal <contact@wearefractal.com> (http://wearefractal.com/)",
-  "main": "./index.js",
+  "main": "./src/index.js",
   "keywords": [
     "gulpplugin"
   ],
